+++ src/main/java/com/wakeUpTogetUp/togetUp/api/mission/service/mapper/ObjectDetectedV32Mapper.java
@@ -1,6 +1,6 @@
-package com.wakeUpTogetUp.togetUp.api.mission.utils.mapper;
+package com.wakeUpTogetUp.togetUp.api.mission.service.mapper;
 
-import com.azure.ai.vision.imageanalysis.DetectedObject;
+import com.microsoft.azure.cognitiveservices.vision.computervision.models.DetectedObject;
 import com.wakeUpTogetUp.togetUp.api.mission.model.BoundingBox;
 import com.wakeUpTogetUp.togetUp.api.mission.model.CustomDetectedObject;
 import java.util.List;
@@ -8,7 +8,8 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ObjectDetectedV40Mapper {
+public class ObjectDetectedV32Mapper {
+
     public List<CustomDetectedObject> toCustomDetectedObjects(List<DetectedObject> objects) {
         return objects.stream()
                 .map(this::toCustomDetectedObject)
@@ -16,15 +17,19 @@ public class ObjectDetectedV40Mapper {
     }
 
     private CustomDetectedObject toCustomDetectedObject(DetectedObject detectedObject) {
+        String parent = detectedObject.parent() != null
+                ? detectedObject.parent().objectProperty()
+                : null;
+
         return CustomDetectedObject.builder()
-                .object(detectedObject.getName())
-                .parent(null)
-                .confidence(detectedObject.getConfidence())
+                .object(detectedObject.objectProperty())
+                .parent(parent)
+                .confidence(detectedObject.confidence())
                 .box(BoundingBox.builder()
-                        .x(detectedObject.getBoundingBox().getX())
-                        .y(detectedObject.getBoundingBox().getY())
-                        .w(detectedObject.getBoundingBox().getW())
-                        .h(detectedObject.getBoundingBox().getH())
+                        .x(detectedObject.rectangle().x())
+                        .y(detectedObject.rectangle().y())
+                        .w(detectedObject.rectangle().w())
+                        .h(detectedObject.rectangle().h())
                         .build())
                 .build();
     }
